Refactor old accounting classes into DTOs.  (#101)
* Refactor old accounting classes into DTOs. Deprecate old accounting CostEstimate and replace it with a DTO akin to the offer DTO. Update DTO description and tests

* Set version of Offer Class correctly

* Add @EqualsAndHashCode back to CostEstimate

* Replace Offer in CostestimateSpec with CostEstimate
